Moved everything into new examples folder
