+++ examples/debug_info.py
