+++ examples/example_pair.py
@@ -13,8 +13,11 @@ running you will be asked to input the 'Security Code' found on
 the back of your IKEA gateway.
 """
 
-import asyncio
-import logging
+# Hack to allow relative import above top level package
+import sys
+import os
+folder = os.path.dirname(os.path.abspath(__file__))  # noqa
+sys.path.insert(0, os.path.normpath("%s/.." % folder))  # noqa
 
 from pytradfri import Gateway
 from pytradfri.command import Command
@@ -23,6 +26,8 @@ from pytradfri.api.aiocoap_api import APIFactory
 from pytradfri.error import PytradfriError
 from pytradfri.util import load_json, save_json
 
+import asyncio
+import logging
 import uuid
 import argparse
